pubsub: move batcher out of internal to allow it be used in external drivers
