+++ internal/website/content/pubsub/batcher/_index.md
@@ -1,0 +1,4 @@
+---
+title: gocloud.dev/pubsub/batcher
+type: pkg
+---
